+++ webapp/e2e-tests/electron/for_feature_development_not_LT_verification/graph_folder_loading/electron-lazy-loading-add-path/types.ts
@@ -1,0 +1,7 @@
+import type { Core as CytoscapeCore } from 'cytoscape';
+import type { ElectronAPI } from '@/shell/electron';
+
+export interface ExtendedWindow {
+  cytoscapeInstance?: CytoscapeCore;
+  electronAPI?: ElectronAPI;
+}
